SET - 2: Extract class - and created a new class called ExtensionFilenameFilter in the utils folder
