+++ src/main/java/de/diddiz/LogBlock/util/ExtensionFilenameFilter.java
@@ -1,0 +1,17 @@
+package de.diddiz.LogBlock.util;
+
+import java.io.File;
+import java.io.FilenameFilter;
+
+public class ExtensionFilenameFilter implements FilenameFilter {
+    private final String ext;
+
+    public ExtensionFilenameFilter(String ext) {
+        this.ext = "." + ext;
+    }
+
+    @Override
+    public boolean accept(File dir, String name) {
+        return name.toLowerCase().endsWith(ext);
+    }
+}
